Add test for editing todo
