+++ src/index.js
@@ -1,8 +1,8 @@
 import './style.css';
 import {
-  addTask, displayTasks, optionsButton, removeButtonClicked,
+  addTask, displayTasks, removeButtonClicked,
 } from './modules/Crud.js';
-import { checkBoxClicked, clearCompletedClicked } from './modules/taskStatusUpdate.js';
+import { checkBoxClicked, optionsButton, clearCompletedClicked } from './modules/taskStatusUpdate.js';
 
 let taskLists = [];
 const taskInput = document.querySelector('.add-task-input');
